Extract functions & style declaration from useCytoscape hook (into themes/ folder)
