+++ src/utils/cytoscape/getMaxEdgeWeight.ts
@@ -1,0 +1,9 @@
+import type { ElementsDefinition } from 'cytoscape';
+
+export function getMaxEdgeWeight(filteredElements: ElementsDefinition) {
+  return (
+    filteredElements?.edges.reduce((max, edge) => {
+      return edge.data.weight > max ? edge.data.weight : max;
+    }, 0) ?? 0
+  );
+}
